requirements.txt
Django==1.9.6
psycopg2==2.6.1
django-oauth-toolkit==0.10.0
djangorestframework==3.3.3
Pillow==3.2.0
django-cors-headers==1.1.0




